feat(ui): Forgot Passcode/Password (#528)
* feat(ui): change recovery seed phrase to shared component

* feat(ui): change set passcode to shared component

* feat(ui): change create password to shared component

* feat(ui): implement forgot auth info

* feat(ui): fix header of forgot auth for android device

* feat(ui): disable backdrop dismiss

* fix(ui): pin verify button on bottom of verify seed phrase page

* fix(ui): fix unit test

* fix(ui): rename biometrics func

* fix(ui): rename biometrics func

---------
