+++ src/store/index.ts
@@ -6,7 +6,7 @@ import { credsCacheSlice } from "./reducers/credsCache";
 import { connectionsCacheSlice } from "./reducers/connectionsCache";
 import { walletConnectionsCacheSlice } from "./reducers/walletConnectionsCache";
 import { identifierViewTypeCacheSlice } from "./reducers/identifierViewTypeCache";
-import { biometryCacheSlice } from "./reducers/biometryCache";
+import { biometricsCacheSlice } from "./reducers/biometricsCache";
 import { credsArchivedCacheSlice } from "./reducers/credsArchivedCache";
 import { ssiAgentSlice } from "./reducers/ssiAgent";
 
@@ -20,7 +20,7 @@ const store = configureStore({
     connectionsCache: connectionsCacheSlice.reducer,
     walletConnectionsCache: walletConnectionsCacheSlice.reducer,
     identifierViewTypeCacheCache: identifierViewTypeCacheSlice.reducer,
-    biometryCache: biometryCacheSlice.reducer,
+    biometricsCache: biometricsCacheSlice.reducer,
     ssiAgentCache: ssiAgentSlice.reducer,
   },
   middleware: (getDefaultMiddleware) =>
