+++ src/store/reducers/biometricsCache/biometricsCache.types.ts
@@ -1,0 +1,4 @@
+interface BiometricsCacheProps {
+  enabled: boolean;
+}
+export type { BiometricsCacheProps };
