--- src/store/reducers/biometryCache/biometryCache.types.ts
@@ -1,4 +1,0 @@
-interface BiometryCacheProps {
-  enabled: boolean;
-}
-export type { BiometryCacheProps };
