--- src/store/reducers/biometryCache/index.ts
@@ -1,1 +1,0 @@
-export * from "./biometryCache";
