+++ src/ui/components/CreatePasscodeModule/CreatePasscodeModule.types.ts
@@ -1,0 +1,13 @@
+interface CreatePasscodeModuleProps {
+  title?: string;
+  description?: string;
+  testId: string;
+  onCreateSuccess: () => void;
+  onPasscodeChange?: (passcode: string, originalPassCode: string) => void;
+}
+
+interface CreatePasscodeModuleRef {
+  clearState: () => void;
+}
+
+export type { CreatePasscodeModuleProps, CreatePasscodeModuleRef };
