+++ src/ui/components/ForgotAuthInfo/ForgotAuthInfo.types.ts
@@ -1,0 +1,13 @@
+enum ForgotType {
+  Passcode = "passcode",
+  Password = "password",
+}
+
+interface ForgotAuthInfoProps {
+  isOpen: boolean;
+  onClose: (shouldCloseParent?: boolean) => void;
+  type: ForgotType;
+}
+
+export type { ForgotAuthInfoProps };
+export { ForgotType };
