+++ src/ui/components/ForgotAuthInfo/index.ts
@@ -1,0 +1,1 @@
+export * from "./ForgotAuthInfo";
