+++ src/ui/components/PasswordModule/PasswordModule.types.ts
@@ -1,0 +1,26 @@
+interface RegexItemProps {
+  condition: boolean;
+  label: string;
+}
+
+interface PasswordRegexProps {
+  password: string;
+}
+
+interface PasswordModuleProps {
+  testId: string;
+  title?: string;
+  description?: string;
+  onCreateSuccess: (skipped: boolean) => void;
+}
+
+interface PasswordModuleRef {
+  clearState: () => void;
+}
+
+export type {
+  PasswordRegexProps,
+  RegexItemProps,
+  PasswordModuleProps,
+  PasswordModuleRef,
+};
