+++ src/ui/components/RecoverySeedPhraseModule/index.ts
@@ -1,0 +1,2 @@
+export * from "./RecoverySeedPhraseModule";
+export * from "./RecoverySeedPhraseModule.types";
