+++ src/ui/pages/LockPage/LockPage.scss
@@ -1,6 +1,5 @@
 .lock-page {
   display: flex !important;
-  z-index: 2147483647; // max value
 
   &.hide {
     .responsive-page-content,
