+++ src/ui/pages/LockPage/LockPage.tsx
@@ -1,46 +1,35 @@
 import i18n from "i18next";
 import { useEffect, useState } from "react";
 import { useSelector } from "react-redux";
-import { ResponsivePageLayout } from "../../components/layout/ResponsivePageLayout";
-import { useAppIonRouter } from "../../hooks";
-import { useAppDispatch, useAppSelector } from "../../../store/hooks";
-import {
-  getAuthentication,
-  login,
-  setAuthentication,
-  setCurrentRoute,
-} from "../../../store/reducers/stateCache";
+import { KeyStoreKeys, SecureStorage } from "../../../core/storage";
+import { useAppDispatch } from "../../../store/hooks";
+import { getBiometricsCacheCache } from "../../../store/reducers/biometricsCache";
+import { login } from "../../../store/reducers/stateCache";
+import { Alert } from "../../components/Alert";
 import {
   ErrorMessage,
   MESSAGE_MILLISECONDS,
 } from "../../components/ErrorMessage";
-import { KeyStoreKeys, SecureStorage } from "../../../core/storage";
-import { RoutePath } from "../../../routes";
-import { PasscodeModule } from "../../components/PasscodeModule";
+import { ForgotAuthInfo } from "../../components/ForgotAuthInfo";
+import { ForgotType } from "../../components/ForgotAuthInfo/ForgotAuthInfo.types";
 import { PageFooter } from "../../components/PageFooter";
-import { Alert } from "../../components/Alert";
-import "./LockPage.scss";
+import { PasscodeModule } from "../../components/PasscodeModule";
+import { ResponsivePageLayout } from "../../components/layout/ResponsivePageLayout";
 import { useBiometricAuth } from "../../hooks/useBiometricsHook";
-import { getBiometryCacheCache } from "../../../store/reducers/biometryCache";
+import "./LockPage.scss";
 
 const LockPage = () => {
   const pageId = "lock-page";
-  const ionRouter = useAppIonRouter();
   const dispatch = useAppDispatch();
-  const authentication = useAppSelector(getAuthentication);
   const [passcode, setPasscode] = useState("");
-  const seedPhrase = authentication.seedPhraseIsSet;
   const [alertIsOpen, setAlertIsOpen] = useState(false);
   const [passcodeIncorrect, setPasscodeIncorrect] = useState(false);
   const { handleBiometricAuth } = useBiometricAuth();
-  const biometryCacheCache = useSelector(getBiometryCacheCache);
+  const biometricsCache = useSelector(getBiometricsCacheCache);
+  const [openRecoveryAuth, setOpenRecoveryAuth] = useState(false);
 
-  const headerText = seedPhrase
-    ? i18n.t("lockpage.alert.text.verify")
-    : i18n.t("lockpage.alert.text.restart");
-  const confirmButtonText = seedPhrase
-    ? i18n.t("lockpage.alert.button.verify")
-    : i18n.t("lockpage.alert.button.restart");
+  const headerText = i18n.t("lockpage.alert.text.verify");
+  const confirmButtonText = i18n.t("lockpage.alert.button.verify");
   const cancelButtonText = i18n.t("lockpage.alert.button.cancel");
 
   const handleClearState = () => {
@@ -59,7 +48,7 @@ const LockPage = () => {
 
   useEffect(() => {
     const runBiometrics = async () => {
-      if (biometryCacheCache.enabled) {
+      if (biometricsCache.enabled) {
         await handleBiometrics();
       }
     };
@@ -109,23 +98,8 @@ const LockPage = () => {
   };
 
   const resetPasscode = () => {
-    SecureStorage.delete(KeyStoreKeys.APP_PASSCODE).then(() => {
-      dispatch(
-        setAuthentication({
-          ...authentication,
-          passcodeIsSet: false,
-        })
-      );
-      dispatch(
-        setCurrentRoute({
-          path: RoutePath.SET_PASSCODE,
-        })
-      );
-      ionRouter.push(RoutePath.SET_PASSCODE, "back", "pop");
-      handleClearState();
-    });
+    setOpenRecoveryAuth(true);
   };
-
   return (
     <ResponsivePageLayout
       pageId={pageId}
@@ -175,6 +149,11 @@ const LockPage = () => {
         confirmButtonText={confirmButtonText}
         cancelButtonText={cancelButtonText}
         actionConfirm={resetPasscode}
+      />
+      <ForgotAuthInfo
+        isOpen={openRecoveryAuth}
+        onClose={() => setOpenRecoveryAuth(false)}
+        type={ForgotType.Passcode}
       />
     </ResponsivePageLayout>
   );
