--- src/ui/pages/VerifyRecoverySeedPhrase/VerifyRecoverySeedPhrase.types.ts
@@ -1,6 +1,0 @@
-interface SeedPhraseInfo {
-  value: string;
-  suggestions: string[];
-}
-
-export type { SeedPhraseInfo };
