+++ src/ui/pages/VerifySeedPhrase/VerifySeedPhrase.scss
@@ -16,7 +16,7 @@
   }
 
   .page-footer {
-    margin-top: auto;
+    margin-top: auto !important;
   }
 
   @media screen and (min-width: 250px) and (max-width: 370px) {
